Fixes #832 Provide a new API for verifying static mock and deprecate old for Mockito
