+++ version.properties
@@ -1,1 +1,1 @@
-version=1.7.0
+version=1.7.1
